(ungetestete) Nutzerauthentifizierungen + getesteter Anmeldungszwang beim erstellen von Thesen

+++ Thesis_Datenbank_Aufgabe/Controllers/FiledController.cs
@@ -59,6 +59,10 @@ namespace Thesis_Datenbank_Aufgabe.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "Id,Title,Description,Bachelor,Master,Status,StudentName,StudentEmail,StudentID,Registration,Filing,Typ,Summary,Strenghts,Weaknesses,Evaluation,ContentVal,LayoutVal,StructureVal,StyleVal,LiteraturVal,DifficultyVal,NoveltyVal,RichnessVal,ContentWt,LayoutWt,StructureWt,StyleWt,LiteratureWt,DifficultyWt,NoveltyWt,RichnessWt,Grade,LastModified")] Thesis thesis)
         {
+            if (System.Web.HttpContext.Current.User.Identity.IsAuthenticated == false)
+            {
+                throw new Exception("Bitte zum erstellen einer Thesis einloggen bzw. registrieren");
+            }
             if (ModelState.IsValid)
             {
                 db.ThesisDb.Add(thesis);
@@ -72,6 +76,10 @@ namespace Thesis_Datenbank_Aufgabe.Controllers
         // GET: Filed/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (Environment.UserDomainName != db.ThesisDb.Find(id).Creator)
+            {
+                throw new Exception("Es dürfen nur selbst angelegte Themen geaendert werden");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -103,6 +111,7 @@ namespace Thesis_Datenbank_Aufgabe.Controllers
         // GET: Filed/Delete/5
         public ActionResult Delete(int? id)
         {
+         
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -120,6 +129,10 @@ namespace Thesis_Datenbank_Aufgabe.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (Environment.UserDomainName != db.ThesisDb.Find(id).Creator)
+            {
+                throw new Exception("Es dürfen nur selbst angelegte Themen geloescht werden");
+            }
             Thesis thesis = db.ThesisDb.Find(id);
             db.ThesisDb.Remove(thesis);
             db.SaveChanges();
