+++ Thesis_Datenbank_Aufgabe/Controllers/OpenController.cs
@@ -46,8 +46,13 @@ namespace Thesis_Datenbank_Aufgabe.Controllers
         // finden Sie unter http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "Id,Title,Description,Bachelor,Master,Status,StudentName,StudentEmail,StudentID,Registration,Filing,Typ,Summary,Strenghts,Weaknesses,Evaluation,ContentVal,LayoutVal,StructureVal,StyleVal,LiteraturVal,DifficultyVal,NoveltyVal,RichnessVal,ContentWt,LayoutWt,StructureWt,StyleWt,LiteratureWt,DifficultyWt,NoveltyWt,RichnessWt,Grade,LastModified")] Thesis thesis)
+        public ActionResult Create([Bind(Include = "Id,Title,Description,Bachelor,Master,Status,StudentName,StudentEmail,StudentID,Registration,Filing,Typ,Summary,Strenghts,Weaknesses,Evaluation,ContentVal,LayoutVal,StructureVal,StyleVal,LiteraturVal,DifficultyVal,NoveltyVal,RichnessVal,ContentWt,LayoutWt,StructureWt,StyleWt,LiteratureWt,DifficultyWt,NoveltyWt,RichnessWt,Grade,LastModified,Creator")] Thesis thesis)
         {
+            //Man muss zum erstellen angemeldet sein
+            if (System.Web.HttpContext.Current.User.Identity.IsAuthenticated==false)
+            {
+                throw new Exception("Bitte zum erstellen einer Thesis einloggen bzw. registrieren");
+            }
             if (ModelState.IsValid)
             {
                 db.ThesisDb.Add(thesis);
@@ -61,6 +66,10 @@ namespace Thesis_Datenbank_Aufgabe.Controllers
         // GET: Open/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (Environment.UserDomainName != db.ThesisDb.Find(id).Creator)
+            {
+                throw new Exception("Es dürfen nur selbst angelegte Themen geaendert werden");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -78,8 +87,10 @@ namespace Thesis_Datenbank_Aufgabe.Controllers
         // finden Sie unter http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "Id,Title,Description,Bachelor,Master,Status,StudentName,StudentEmail,StudentID,Registration,Filing,Typ,Summary,Strenghts,Weaknesses,Evaluation,ContentVal,LayoutVal,StructureVal,StyleVal,LiteraturVal,DifficultyVal,NoveltyVal,RichnessVal,ContentWt,LayoutWt,StructureWt,StyleWt,LiteratureWt,DifficultyWt,NoveltyWt,RichnessWt,Grade,LastModified")] Thesis thesis)
+        public ActionResult Edit([Bind(Include = "Id,Title,Description,Bachelor,Master,Status,StudentName,StudentEmail,StudentID,Registration,Filing,Typ,Summary,Strenghts,Weaknesses,Evaluation,ContentVal,LayoutVal,StructureVal,StyleVal,LiteraturVal,DifficultyVal,NoveltyVal,RichnessVal,ContentWt,LayoutWt,StructureWt,StyleWt,LiteratureWt,DifficultyWt,NoveltyWt,RichnessWt,Grade,LastModified,Creator")] Thesis thesis)
         {
+     
+             
             if (ModelState.IsValid)
             {
                 db.Entry(thesis).State = EntityState.Modified;
@@ -92,6 +103,7 @@ namespace Thesis_Datenbank_Aufgabe.Controllers
         // GET: Open/Delete/5
         public ActionResult Delete(int? id)
         {
+           
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -109,6 +121,10 @@ namespace Thesis_Datenbank_Aufgabe.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (Environment.UserDomainName != db.ThesisDb.Find(id).Creator)
+            {
+                throw new Exception("Es dürfen nur selbst angelegte Themen geloescht werden");
+            }
             Thesis thesis = db.ThesisDb.Find(id);
             db.ThesisDb.Remove(thesis);
             db.SaveChanges();
